change type on rankChange from u8 to int

+++ Source/Core/Core/Slippi/SlippiUser.cpp
@@ -236,7 +236,7 @@ SlippiUser::RankInfo SlippiUser::FetchUserRank(std::string connectCode)
 	float ratingChange = (userRank.ratingOrdinal > 0.001f) ? ratingOrdinal - userRank.ratingOrdinal : 0;
 	INFO_LOG(SLIPPI_ONLINE, "Rating Change: %0.1f", ratingChange);
 
-	u8 rankChange = (userRank.rank > 0.001f) ? rank - userRank.rank: 0;
+	int rankChange = (userRank.rank > 0.001f) ? rank - userRank.rank : 0;
 	INFO_LOG(SLIPPI_ONLINE, "Rank Change: %d", rankChange);
 
 	info.ratingOrdinal = ratingOrdinal;
